Move user store to modules
