+++ packages/client-web/src/store/index.ts
@@ -2,7 +2,7 @@ import { createStore, applyMiddleware, combineReducers } from "redux";
 import thunkMiddleware from "redux-thunk";
 import { composeWithDevTools } from "redux-devtools-extension";
 
-import user from "./user";
+import user from "../modules/user/model/store";
 import date from "../modules/date/model/store";
 import feeds from "../modules/feeds/model/store";
 import timeline from "../modules/timeline/model/store";
